requirements.txt
django~=5.2.0
pillow
django-environ
pyhub-llm[all]